+++ libnavui-dropin/src/test/java/com/mapbox/navigation/dropin/infopanel/InfoPanelHeaderArrivalBinderTest.kt
@@ -18,6 +18,7 @@ import com.mapbox.navigation.testing.MainCoroutineRule
 import io.mockk.mockk
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Assert.assertNotNull
+import org.junit.Assert.assertNull
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -77,5 +78,17 @@ class InfoPanelHeaderArrivalBinderTest {
         components.onAttached(mapboxNavigation)
 
         assertNotNull(components.findComponent { it is EndNavigationButtonComponent })
+    }
+
+    @Test
+    @Suppress("MaxLineLength")
+    fun `bind should NOT bind EndNavigationButtonComponent when ViewOptionsCustomization showEndNavigationButton is FALSE`() {
+        navContext.applyOptionsCustomization {
+            showEndNavigationButton = false
+        }
+        val components = sut.bind(FrameLayout(ctx))
+        components.onAttached(mapboxNavigation)
+
+        assertNull(components.findComponent { it is EndNavigationButtonComponent })
     }
 }