+++ libnavui-dropin/src/test/java/com/mapbox/navigation/dropin/maneuver/ManeuverCoordinatorTest.kt
@@ -1,44 +1,65 @@
 package com.mapbox.navigation.dropin.maneuver
 
+import android.content.Context
 import android.view.ViewGroup
+import android.widget.FrameLayout
+import androidx.test.core.app.ApplicationProvider
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.dropin.EmptyBinder
 import com.mapbox.navigation.dropin.navigationview.NavigationViewContext
+import com.mapbox.navigation.dropin.navigationview.NavigationViewModel
+import com.mapbox.navigation.dropin.testutil.TestLifecycleOwner
 import com.mapbox.navigation.dropin.util.TestStore
+import com.mapbox.navigation.testing.LoggingFrontendTestRule
+import com.mapbox.navigation.testing.MainCoroutineRule
 import com.mapbox.navigation.ui.app.internal.State
 import com.mapbox.navigation.ui.app.internal.navigation.NavigationState
 import com.mapbox.navigation.ui.base.lifecycle.Binder
 import com.mapbox.navigation.ui.base.lifecycle.UIBinder
-import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.take
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runBlockingTest
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
+import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
 
 @OptIn(ExperimentalPreviewMapboxNavigationAPI::class, ExperimentalCoroutinesApi::class)
+@RunWith(RobolectricTestRunner::class)
 class ManeuverCoordinatorTest {
 
-    private val mapboxNavigation = mockk<MapboxNavigation>()
-    private val store = TestStore()
-    private val maneuverFlow = MutableStateFlow<UIBinder?>(null)
-    private val context = mockk<NavigationViewContext> {
-        every { store } returns this@ManeuverCoordinatorTest.store
-        every { uiBinders } returns mockk {
-            every { maneuver } returns maneuverFlow
-        }
-        every { mapStyleLoader } returns mockk(relaxed = true)
-        every { options } returns mockk(relaxed = true)
-        every { styles } returns mockk(relaxed = true)
+    @get:Rule
+    val loggerRule = LoggingFrontendTestRule()
+
+    @get:Rule
+    val coroutineRule = MainCoroutineRule()
+
+    private lateinit var mapboxNavigation: MapboxNavigation
+    private lateinit var store: TestStore
+    private lateinit var context: NavigationViewContext
+    private lateinit var coordinator: ManeuverCoordinator
+
+    @Before
+    fun setUp() {
+        val ctx: Context = ApplicationProvider.getApplicationContext()
+        store = TestStore()
+        context = NavigationViewContext(
+            context = ctx,
+            lifecycleOwner = TestLifecycleOwner(),
+            viewModel = NavigationViewModel(),
+            storeProvider = { store }
+        )
+        mapboxNavigation = mockk(relaxed = true)
+
+        coordinator = ManeuverCoordinator(context, FrameLayout(ctx))
     }
-    private val coordinator = ManeuverCoordinator(context, mockk())
 
     @Test
     fun `should create empty binder fore FreeDrive`() = runBlockingTest {
@@ -52,7 +73,9 @@ class ManeuverCoordinatorTest {
     @Test
     fun `should create empty binder fore FreeDrive even with custom binder`() = runBlockingTest {
         val customBinder = mockk<UIBinder>()
-        maneuverFlow.value = customBinder
+        context.applyBinderCustomization {
+            maneuverBinder = customBinder
+        }
         store.setState(State(navigation = NavigationState.FreeDrive))
         coordinator.apply {
             val binders = mapboxNavigation.flowViewBinders().take(1).toList()
@@ -99,7 +122,9 @@ class ManeuverCoordinatorTest {
     @Test
     fun `should use custom binder for ActiveGuidance`() = runBlockingTest {
         val customBinder = mockk<UIBinder>()
-        maneuverFlow.value = customBinder
+        context.applyBinderCustomization {
+            maneuverBinder = customBinder
+        }
         store.setState(State(navigation = NavigationState.ActiveNavigation))
         coordinator.apply {
             val binders = mapboxNavigation.flowViewBinders().take(1).toList()
@@ -123,57 +148,17 @@ class ManeuverCoordinatorTest {
     }
 
     @Test
-    fun `should create new binder when custom binders change`() = runBlockingTest {
-        val customBinder1 = mockk<UIBinder>()
-        val customBinder2 = mockk<UIBinder>()
-        val collectedBinders = mutableListOf<Binder<ViewGroup>>()
-        store.setState(State(navigation = NavigationState.ActiveNavigation))
-        coordinator.apply {
-            val job = launch {
-                mapboxNavigation.flowViewBinders().take(4).toList(collectedBinders)
+    fun `should return EmptyBinder when ViewOptionsCustomization showManeuver is FALSE`() =
+        runBlockingTest {
+            context.applyOptionsCustomization {
+                showManeuver = false
             }
-            maneuverFlow.value = customBinder1
-            maneuverFlow.value = null
-            maneuverFlow.value = customBinder2
-            job.join()
-            assertEquals(4, collectedBinders.size)
-            assertTrue(collectedBinders[0] is ManeuverViewBinder)
-            assertTrue(collectedBinders[1] === customBinder1)
-            assertTrue(collectedBinders[2] is ManeuverViewBinder)
-            assertTrue(collectedBinders[3] === customBinder2)
-        }
-    }
 
-    @Test
-    fun `should use different instances of ManeuverViewBinder`() = runBlockingTest {
-        val collectedBinders = mutableListOf<Binder<ViewGroup>>()
-        store.setState(State(navigation = NavigationState.ActiveNavigation))
-        coordinator.apply {
-            val job = launch {
-                mapboxNavigation.flowViewBinders().take(3).toList(collectedBinders)
+            val binders = coordinator.run {
+                val mapboxNavigation = mockk<MapboxNavigation>()
+                mapboxNavigation.flowViewBinders().take(1).toList()
             }
-            maneuverFlow.value = mockk()
-            maneuverFlow.value = null
-            job.join()
-            assertTrue(collectedBinders[0] is ManeuverViewBinder)
-            assertTrue(collectedBinders[2] is ManeuverViewBinder)
-            assertFalse(collectedBinders[0] === collectedBinders[2])
-        }
-    }
 
-    @Test
-    fun `should use different instances of EmptyBinder`() = runBlockingTest {
-        val collectedBinders = mutableListOf<Binder<ViewGroup>>()
-        coordinator.apply {
-            val job = launch {
-                mapboxNavigation.flowViewBinders().take(3).toList(collectedBinders)
-            }
-            store.setState(State(navigation = NavigationState.ActiveNavigation))
-            store.setState(State(navigation = NavigationState.FreeDrive))
-            job.join()
-            assertTrue(collectedBinders[0] is EmptyBinder)
-            assertTrue(collectedBinders[2] is EmptyBinder)
-            assertFalse(collectedBinders[0] === collectedBinders[2])
+            assertTrue(binders.first() is EmptyBinder)
         }
-    }
 }